fix: logic related expect/actual classes modality
