+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/configuration/ClientConfigImpl.kt
@@ -25,7 +25,7 @@ import com.wire.kalium.logic.data.client.DeviceType
 import android.provider.Settings
 import com.wire.kalium.logic.data.client.ClientType
 
-actual class ClientConfigImpl(private val context: Context) : ClientConfig {
+actual open class ClientConfigImpl(private val context: Context) : ClientConfig {
 
     actual override fun deviceType(): DeviceType {
         val screenSize = context.resources.configuration.screenLayout and Configuration.SCREENLAYOUT_SIZE_MASK
