+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/call/FlowManagerServiceImpl.kt
@@ -33,7 +33,7 @@ import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.withContext
 
-actual class FlowManagerServiceImpl(
+actual open class FlowManagerServiceImpl(
     appContext: PlatformContext,
     scope: CoroutineScope,
     private val dispatchers: KaliumDispatcher = KaliumDispatcherImpl
