+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/rootDetection/RootDetectorImpl.kt
@@ -3,7 +3,7 @@ package com.wire.kalium.logic.feature.rootDetection
 import android.annotation.SuppressLint
 import java.io.File
 
-actual class RootDetectorImpl actual constructor() : RootDetector {
+actual open class RootDetectorImpl actual constructor() : RootDetector {
 
     actual override fun isSystemRooted(): Boolean {
         // Lambdas so they can be executed lazily
