+++ logic/src/appleMain/kotlin/com/wire/kalium/logic/configuration/ClientConfigImpl.kt
@@ -30,7 +30,7 @@ import kotlinx.cinterop.value
 import platform.Foundation.NSProcessInfo
 import platform.darwin.sysctlbyname
 
-actual class ClientConfigImpl : ClientConfig {
+actual open class ClientConfigImpl : ClientConfig {
     actual override fun deviceType(): DeviceType {
         // TODO: Figure out the actual darwin device type
         return DeviceType.Desktop
