+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/configuration/ClientConfigImpl.kt
@@ -20,17 +20,8 @@ package com.wire.kalium.logic.configuration
 
 import com.wire.kalium.logic.data.client.ClientType
 import com.wire.kalium.logic.data.client.DeviceType
-import io.mockative.Mockable
 
-@Mockable
-interface ClientConfig {
-    fun deviceType(): DeviceType
-    fun deviceModelName(): String
-    fun deviceName(): String
-    fun clientType(): ClientType
-}
-
-expect class ClientConfigImpl : ClientConfig {
+expect open class ClientConfigImpl : ClientConfig {
     override fun deviceType(): DeviceType
     override fun deviceModelName(): String
     override fun deviceName(): String
