+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/sync/WorkSchedulerImpl.kt
@@ -51,7 +51,7 @@ internal actual class GlobalWorkSchedulerImpl(
     }
 }
 
-internal actual class UserSessionWorkSchedulerImpl(
+internal actual open class UserSessionWorkSchedulerImpl(
     actual override val scope: UserSessionScope,
 ) : UserSessionWorkScheduler {
